Refactored include paths, enhanced directory handling, adjusted CMake configuration, and optimized model loading logic.
